+++ bluetooth_low_energy_platform_interface/lib/src/bluetooth_low_energy_manager.dart
@@ -5,12 +5,12 @@ import 'bluetooth_low_energy_state.dart';
 
 /// The abstract base class that manages central and peripheral objects.
 abstract class BluetoothLowEnergyManager implements LogController {
-  /// The current state of the manager.
-  BluetoothLowEnergyState get state;
-
-  /// Tells the manager’s state updated.
+  /// Tells the manager's state updated.
   Stream<BluetoothLowEnergyStateChangedEventArgs> get stateChanged;
 
   /// Sets up the manager.
   Future<void> setUp();
+
+  /// Gets the manager's state.
+  Future<BluetoothLowEnergyState> getState();
 }