+++ bluetooth_low_energy_platform_interface/lib/src/my_gatt_characteristic.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'gatt_characteristic.dart';
 import 'gatt_characteristic_property.dart';
 import 'my_gatt_attribute.dart';
@@ -7,12 +9,14 @@ class MyGattCharacteristic extends MyGattAttribute
     implements GattCharacteristic {
   @override
   final List<GattCharacteristicProperty> properties;
+  final Uint8List? value;
   @override
   final List<MyGattDescriptor> descriptors;
 
   MyGattCharacteristic({
     required super.uuid,
     required this.properties,
+    this.value,
     required this.descriptors,
   });
 }